requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
scikit-learn>=1.1.0
statsmodels>=0.13.0
ete3>=3.1.0
pyyaml>=6.0
threadpoolctl>=3.1.0
biopython>=1.79
